pyproject.toml
[project]
name = "smartnews-pipeline"
version = "0.1.0"
description = "SmartNews data pipeline — RSS ingestion, AI enrichment, DuckDB storage"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "duckdb>=0.10",
    "openai>=1.0",
    "trafilatura>=1.8",
    "lxml_html_clean>=0.1",
    "scikit-learn>=1.3",
    "pydantic>=2.5",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mcp-server-duckdb>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
